SHAndwich PoW hash function & new genesis blocks
* Add SHAndwich hasher

* Copy Litecoin: add seperate GetPowHash() function to block header
class

* Copy Litecoin: disable PoW check when loading blocks we have saved to
our own hdd. Otherwise, the SHAndwich PoW would have to be re-calculated
every time a block is loaded from disk as it is only used for PoW and
then forgotten.

(See comments in code for credit to thrasher- and shaolinfry who are
credited with those ideas by git blame)

* Replace SHA256D with SHAndwich for PoW

* Update chainparams with new genesis block

* Removed invalid Bitcoin Core seed nodes from chainparams

* Updated chainparams block checkpoints

* Updated some unit tests. TODO miner_tests block data struct needs to
be updated (disabled for now)